// frontend/src/stores/types.ts
export interface ETFUniverse {
    STOCKS: string[]
    BONDS: string[]
    COMMODITIES: string[]
    ALTERNATIVES: string[]
}

export interface EnabledCategories {
    STOCKS: boolean
    BONDS: boolean
    COMMODITIES: boolean
    ALTERNATIVES: boolean
}